split at.js config from comment.js
